requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
werkzeug==2.0.1
openai==1.12.0
pyephem==4.1.5 